render state refactor
