+++ packages/engine/src/renderer/WebGLRendererSystem.ts
@@ -53,12 +53,8 @@ import { defaultPostProcessingSchema } from '../scene/constants/PostProcessing'
 import { createWebXRManager, WebXRManager } from '../xr/WebXRManager'
 import { isHeadset, useIsHeadset, XRState } from '../xr/XRState'
 import { LinearTosRGBEffect } from './effects/LinearTosRGBEffect'
-import {
-  accessEngineRendererState,
-  EngineRendererAction,
-  EngineRendererReceptor,
-  EngineRendererState
-} from './EngineRendererState'
+import { accessEngineRendererState, EngineRendererAction, EngineRendererState } from './EngineRendererState'
+import { changeRenderMode } from './functions/changeRenderMode'
 import { configureEffectComposer } from './functions/configureEffectComposer'
 import { updateShadowMap } from './functions/RenderSettingsFunction'
 import WebGL from './THREE.WebGL'
@@ -346,9 +342,11 @@ export default async function WebGLRendererSystem(world: World) {
     default: DefaultPostProcessingState
   }
 
+  const engineRendererState = getState(EngineRendererState)
+
   const reactor = startReactor(() => {
     const renderSettings = useHookstate(getRendererSceneMetadataState(world))
-    const engineRendererSettings = useHookstate(getState(EngineRendererState))
+    const engineRendererSettings = useHookstate(engineRendererState)
     const postprocessing = useHookstate(getPostProcessingSceneMetadataState(world))
 
     useEffect(() => {
@@ -388,16 +386,61 @@ export default async function WebGLRendererSystem(world: World) {
     return null
   })
 
+  function setQualityLevel(qualityLevel) {
+    EngineRenderer.instance.scaleFactor = qualityLevel / EngineRenderer.instance.maxQualityLevel
+    EngineRenderer.instance.renderer.setPixelRatio(window.devicePixelRatio * EngineRenderer.instance.scaleFactor)
+    EngineRenderer.instance.needsResize = true
+  }
+
+  function setUsePostProcessing(usePostProcessing) {
+    if (engineRendererState.usePostProcessing.value === usePostProcessing) return
+    usePostProcessing = EngineRenderer.instance.supportWebGL2 && usePostProcessing
+    configureEffectComposer(!usePostProcessing)
+  }
+
   const execute = () => {
-    for (const action of setQualityLevelActions()) EngineRendererReceptor.setQualityLevel(action)
-    for (const action of setAutomaticActions()) EngineRendererReceptor.setAutomatic(action)
-    for (const action of setPostProcessingActions()) EngineRendererReceptor.setPostProcessing(action)
-    for (const action of setShadowsActions()) EngineRendererReceptor.setShadows(action)
-    for (const action of setDebugActions()) EngineRendererReceptor.setDebug(action)
-    for (const action of changedRenderModeActions()) EngineRendererReceptor.changedRenderMode(action)
-    for (const action of changeNodeHelperVisibilityActions()) EngineRendererReceptor.changeNodeHelperVisibility(action)
-    for (const action of changeGridToolHeightActions()) EngineRendererReceptor.changeGridToolHeight(action)
-    for (const action of changeGridToolVisibilityActions()) EngineRendererReceptor.changeGridToolVisibility(action)
+    for (const action of setQualityLevelActions()) {
+      const s = engineRendererState
+      s.qualityLevel.set(action.qualityLevel)
+      setQualityLevel(action.qualityLevel)
+    }
+    for (const action of setAutomaticActions()) {
+      const s = engineRendererState
+      s.automatic.set(action.automatic)
+    }
+    for (const action of setPostProcessingActions()) {
+      if (action.usePostProcessing && isHeadset()) return
+      setUsePostProcessing(action.usePostProcessing)
+      const s = engineRendererState
+      s.usePostProcessing.set(action.usePostProcessing)
+    }
+    for (const action of setShadowsActions()) {
+      if (action.useShadows && isHeadset()) return
+      const s = getState(EngineRendererState)
+      s.useShadows.set(action.useShadows)
+      if (!Engine.instance.isEditor) updateShadowMap()
+    }
+    for (const action of setDebugActions()) {
+      const s = getState(EngineRendererState)
+      s.debugEnable.set(action.debugEnable)
+    }
+    for (const action of changedRenderModeActions()) {
+      changeRenderMode(action.renderMode)
+      const s = getState(EngineRendererState)
+      s.renderMode.set(action.renderMode)
+    }
+    for (const action of changeNodeHelperVisibilityActions()) {
+      const s = getState(EngineRendererState)
+      s.nodeHelperVisibility.set(action.visibility)
+    }
+    for (const action of changeGridToolHeightActions()) {
+      const s = getState(EngineRendererState)
+      s.gridHeight.set(action.gridHeight)
+    }
+    for (const action of changeGridToolVisibilityActions()) {
+      const s = getState(EngineRendererState)
+      s.gridVisibility.set(action.visibility)
+    }
 
     EngineRenderer.instance.execute(world.deltaSeconds)
   }
